Fixed issues on build
Issue : Initialization scene was loading the LoadEventChannleSO and SceneLoader on PersistantManagers scene wasn't able to communicate with loaded one. Because both of the scripts had their own references. Issue fixed by using an UnityAction<LoadEventChannelSO> to let the SceneLoader script know when LoadEventChannelSO loaded and refresh its null references with that. On the other hand still dont know what is behind this problem.